Added Login stuff (not tested)
when you go to the login page, if logged in, it will log out (clear session data). If post made to server, it will check if username exists and add if not. Then gets the user_id and saves to session variable.
User sees basic login form with skip login link when logging in.
Header has adjustments to hide in-accessible options and add login page link.